+++ lib/v2/njust/dgxg.js
@@ -1,17 +1,18 @@
 const cheerio = require('cheerio');
 const { parseDate } = require('@/utils/parse-date');
 const timezone = require('@/utils/timezone');
-const { getContent } = require('../eo/util');
+const { getContent } = require('./utils');
 
 const map = new Map([
-    ['tzgg', { title: '南京理工大学电子工程与光电技术学院 -- 通知公告', id: '/1920' }],
-    ['xwdt', { title: '南京理工大学电子工程与光电技术学院 -- 新闻动态', id: '/1919' }],
+    ['gstz', { title: '南京理工大学电光学院研学网 -- 公示通知', id: '/6509' }],
+    ['xswh', { title: '南京理工大学电光学院研学网 -- 学术文化', id: '/6511' }],
+    ['jyzd', { title: '南京理工大学电光学院研学网 -- 就业指导', id: '/6510' }],
 ]);
 
-const host = 'https://eoe.njust.edu.cn';
+const host = 'https://dgxg.njust.edu.cn';
 
 module.exports = async (ctx) => {
-    const type = ctx.params.type || 'tzgg';
+    const type = ctx.params.type || 'gstz';
     const info = map.get(type);
     if (!info) {
         throw 'invalid type';
@@ -21,7 +22,7 @@ module.exports = async (ctx) => {
 
     const html = await getContent(siteUrl, true);
     const $ = cheerio.load(html);
-    const list = $('ul.news_ul').find('li');
+    const list = $('ul.wp_article_list').find('li');
 
     ctx.state.data = {
         title: info.title,
@@ -31,7 +32,7 @@ module.exports = async (ctx) => {
             list
                 .map((index, item) => ({
                     title: $(item).find('a').attr('title').trim(),
-                    pubDate: timezone(parseDate($(item).find('span').text(), 'YYYY-MM-DD'), +8),
+                    pubDate: timezone(parseDate($(item).find('span.Article_PublishDate').text(), 'YYYY-MM-DD'), +8),
                     link: $(item).find('a').attr('href'),
                 }))
                 .get(),
