+++ lib/v2/njust/eo.js
@@ -1,7 +1,7 @@
 const cheerio = require('cheerio');
 const { parseDate } = require('@/utils/parse-date');
 const timezone = require('@/utils/timezone');
-const { getContent } = require('../eo/util');
+const { getContent } = require('./utils');
 
 const map = new Map([
     ['16tz', { title: '南京理工大学电光16 -- 通知公告', id: '/_t217/tzgg' }],
