+++ lib/v2/njust/eoe.js
@@ -1,19 +1,17 @@
 const cheerio = require('cheerio');
 const { parseDate } = require('@/utils/parse-date');
 const timezone = require('@/utils/timezone');
-const { getContent } = require('../eo/util');
+const { getContent } = require('./utils');
 
 const map = new Map([
-    ['jstz', { title: '南京理工大学教务处 -- 教师通知', id: '/1216' }],
-    ['xstz', { title: '南京理工大学教务处 -- 学生通知', id: '/1217' }],
-    ['xw', { title: '南京理工大学教务处 -- 新闻', id: '/1218' }],
-    ['xydt', { title: '南京理工大学教务处 -- 学院动态', id: '/1219' }],
+    ['tzgg', { title: '南京理工大学电子工程与光电技术学院 -- 通知公告', id: '/1920' }],
+    ['xwdt', { title: '南京理工大学电子工程与光电技术学院 -- 新闻动态', id: '/1919' }],
 ]);
 
-const host = 'https://jwc.njust.edu.cn';
+const host = 'https://eoe.njust.edu.cn';
 
 module.exports = async (ctx) => {
-    const type = ctx.params.type || 'xstz';
+    const type = ctx.params.type || 'tzgg';
     const info = map.get(type);
     if (!info) {
         throw 'invalid type';
@@ -23,7 +21,7 @@ module.exports = async (ctx) => {
 
     const html = await getContent(siteUrl, true);
     const $ = cheerio.load(html);
-    const list = $('div#wp_news_w3').find('tr');
+    const list = $('ul.news_ul').find('li');
 
     ctx.state.data = {
         title: info.title,
@@ -33,7 +31,7 @@ module.exports = async (ctx) => {
             list
                 .map((index, item) => ({
                     title: $(item).find('a').attr('title').trim(),
-                    pubDate: timezone(parseDate($(item).find('td[width="14%"]').text(), 'YYYY-MM-DD'), +8),
+                    pubDate: timezone(parseDate($(item).find('span').text(), 'YYYY-MM-DD'), +8),
                     link: $(item).find('a').attr('href'),
                 }))
                 .get(),
