+++ lib/v2/njust/gs.js
@@ -1,7 +1,7 @@
 const cheerio = require('cheerio');
 const { parseDate } = require('@/utils/parse-date');
 const timezone = require('@/utils/timezone');
-const { getContent } = require('../eo/util');
+const { getContent } = require('./utils');
 
 const map = new Map([
     ['tzgg', { title: '南京理工大学研究生院 -- 通知公告', id: '/sytzgg_4568' }],
