+++ lib/v2/njust/jwc.js
@@ -1,18 +1,19 @@
 const cheerio = require('cheerio');
 const { parseDate } = require('@/utils/parse-date');
 const timezone = require('@/utils/timezone');
-const { getContent } = require('../eo/util');
+const { getContent } = require('./utils');
 
 const map = new Map([
-    ['gstz', { title: '南京理工大学电光学院研学网 -- 公示通知', id: '/6509' }],
-    ['xswh', { title: '南京理工大学电光学院研学网 -- 学术文化', id: '/6511' }],
-    ['jyzd', { title: '南京理工大学电光学院研学网 -- 就业指导', id: '/6510' }],
+    ['jstz', { title: '南京理工大学教务处 -- 教师通知', id: '/1216' }],
+    ['xstz', { title: '南京理工大学教务处 -- 学生通知', id: '/1217' }],
+    ['xw', { title: '南京理工大学教务处 -- 新闻', id: '/1218' }],
+    ['xydt', { title: '南京理工大学教务处 -- 学院动态', id: '/1219' }],
 ]);
 
-const host = 'https://dgxg.njust.edu.cn';
+const host = 'https://jwc.njust.edu.cn';
 
 module.exports = async (ctx) => {
-    const type = ctx.params.type || 'gstz';
+    const type = ctx.params.type || 'xstz';
     const info = map.get(type);
     if (!info) {
         throw 'invalid type';
@@ -22,7 +23,7 @@ module.exports = async (ctx) => {
 
     const html = await getContent(siteUrl, true);
     const $ = cheerio.load(html);
-    const list = $('ul.wp_article_list').find('li');
+    const list = $('div#wp_news_w3').find('tr');
 
     ctx.state.data = {
         title: info.title,
@@ -32,7 +33,7 @@ module.exports = async (ctx) => {
             list
                 .map((index, item) => ({
                     title: $(item).find('a').attr('title').trim(),
-                    pubDate: timezone(parseDate($(item).find('span.Article_PublishDate').text(), 'YYYY-MM-DD'), +8),
+                    pubDate: timezone(parseDate($(item).find('td[width="14%"]').text(), 'YYYY-MM-DD'), +8),
                     link: $(item).find('a').attr('href'),
                 }))
                 .get(),
